+++ src/components/Molecules/HeaderEsuWithIcons/HeaderEsuWithIcons.md
@@ -1,0 +1,14 @@
+```js
+<div>
+  <HeaderEsuWithIcons
+    link={'/search'}
+    iconType="Search"
+    title="Click here to search"
+  />
+  <HeaderEsuWithIcons
+    link={'https://www.facebook.com'}
+    iconType="Email"
+    title="Sign up for our emails"
+  />
+</div>
+```
