+++ src/components/Molecules/HeaderEsuWithIcons/assets/HeaderIcons.js
