--- src/components/Molecules/HeaderIcon/HeaderIcon.md
@@ -1,10 +1,0 @@
-```js
-<div>
-  <HeaderIcon link={'/search'} iconType="Search" title="Click here to search" />
-  <HeaderIcon
-    link={'https://www.facebook.com'}
-    iconType="Email"
-    title="Sign up for our emails"
-  />
-</div>
-```
